클라이언트 부분 분리, 윤제형 계정 붙이기 시도중. hwnd 삭제한걸로 교체..
