+++ Acquitaine Engine/Acquitaine Client/ClientEntry.h
@@ -1,0 +1,6 @@
+#pragma once
+class ClientEntry
+{
+
+};
+
