+++ Acquitaine Engine/Acquitaine Client/SceneManager.h
@@ -1,0 +1,13 @@
+#pragma once
+
+class ParentScene;
+class TestScene;
+
+class SceneManager
+{
+public:
+	SceneManager();
+	~SceneManager();
+	void Initialize();
+};
+
