+++ Acquitaine Engine/Acquitaine Client/TestScene.h
@@ -1,0 +1,12 @@
+#pragma once
+#include "ParentScene.h"
+
+class TestObject;
+
+class TestScene : public act::ParentScene
+{
+public:
+	TestScene(string SceneName);
+	virtual ~TestScene();
+};
+
