+++ Acquitaine Engine/Acquitaine Engine.sln
@@ -4,6 +4,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 17.3.32825.248
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Acquitaine Engine", "Acquitaine Engine\Acquitaine Engine.vcxproj", "{89838E02-4962-4DE6-8968-DB79F031BAD5}"
+	ProjectSection(ProjectDependencies) = postProject
+		{ED4100F7-F193-4458-9BC9-CD239BFF1389} = {ED4100F7-F193-4458-9BC9-CD239BFF1389}
+	EndProjectSection
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Acquitaine Game Engine", "Acquitaine Game Engine\Acquitaine Game Engine.vcxproj", "{4D7CDE76-B2EC-4BD2-90CF-49E1EC46714B}"
 EndProject
@@ -12,6 +15,13 @@ EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AptoMath", "..\Apto\AptoMath\Apto Math.vcxproj", "{AD83A6C0-1A44-4522-8DB9-E88A6805F670}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AptoUtil", "..\Apto\AptoUtil\Apto Util.vcxproj", "{215CC7A8-FBE2-4A97-BBC4-DE25DE5E2D15}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Acquitaine Game Client", "Acquitaine Client\Acquitaine Client.vcxproj", "{ED4100F7-F193-4458-9BC9-CD239BFF1389}"
+	ProjectSection(ProjectDependencies) = postProject
+		{4D7CDE76-B2EC-4BD2-90CF-49E1EC46714B} = {4D7CDE76-B2EC-4BD2-90CF-49E1EC46714B}
+	EndProjectSection
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DXRenderer", "..\YunGine\DXRenderer\DXRenderer.vcxproj", "{87EFCDB8-583E-47B8-8EA1-675326BE8A38}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -61,6 +71,22 @@ Global
 		{215CC7A8-FBE2-4A97-BBC4-DE25DE5E2D15}.Release|x64.Build.0 = Release|x64
 		{215CC7A8-FBE2-4A97-BBC4-DE25DE5E2D15}.Release|x86.ActiveCfg = Release|Win32
 		{215CC7A8-FBE2-4A97-BBC4-DE25DE5E2D15}.Release|x86.Build.0 = Release|Win32
+		{ED4100F7-F193-4458-9BC9-CD239BFF1389}.Debug|x64.ActiveCfg = Debug|x64
+		{ED4100F7-F193-4458-9BC9-CD239BFF1389}.Debug|x64.Build.0 = Debug|x64
+		{ED4100F7-F193-4458-9BC9-CD239BFF1389}.Debug|x86.ActiveCfg = Debug|Win32
+		{ED4100F7-F193-4458-9BC9-CD239BFF1389}.Debug|x86.Build.0 = Debug|Win32
+		{ED4100F7-F193-4458-9BC9-CD239BFF1389}.Release|x64.ActiveCfg = Release|x64
+		{ED4100F7-F193-4458-9BC9-CD239BFF1389}.Release|x64.Build.0 = Release|x64
+		{ED4100F7-F193-4458-9BC9-CD239BFF1389}.Release|x86.ActiveCfg = Release|Win32
+		{ED4100F7-F193-4458-9BC9-CD239BFF1389}.Release|x86.Build.0 = Release|Win32
+		{87EFCDB8-583E-47B8-8EA1-675326BE8A38}.Debug|x64.ActiveCfg = Debug|x64
+		{87EFCDB8-583E-47B8-8EA1-675326BE8A38}.Debug|x64.Build.0 = Debug|x64
+		{87EFCDB8-583E-47B8-8EA1-675326BE8A38}.Debug|x86.ActiveCfg = Debug|Win32
+		{87EFCDB8-583E-47B8-8EA1-675326BE8A38}.Debug|x86.Build.0 = Debug|Win32
+		{87EFCDB8-583E-47B8-8EA1-675326BE8A38}.Release|x64.ActiveCfg = Release|x64
+		{87EFCDB8-583E-47B8-8EA1-675326BE8A38}.Release|x64.Build.0 = Release|x64
+		{87EFCDB8-583E-47B8-8EA1-675326BE8A38}.Release|x86.ActiveCfg = Release|Win32
+		{87EFCDB8-583E-47B8-8EA1-675326BE8A38}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
