+++ Acquitaine Engine/Acquitaine Game Engine/API.cpp
@@ -1,0 +1,6 @@
+#include "API.h"
+
+namespace act
+{
+
+}
