+++ Acquitaine Engine/Acquitaine Game Engine/API.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace act
+{
+
+}
