+++ Acquitaine Engine/Acquitaine Game Engine/Acquitaine Game Engine.vcxproj.filters
@@ -13,15 +13,6 @@
     <Filter Include="Scene">
       <UniqueIdentifier>{def1847b-72f6-48b6-a8aa-bb6a3127e2fc}</UniqueIdentifier>
     </Filter>
-    <Filter Include="TestSpace">
-      <UniqueIdentifier>{cdf1ab62-61c8-44f2-8cd8-2fc983a658c1}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="TestSpace\Object">
-      <UniqueIdentifier>{fba7ce2f-2986-4c51-98be-1a1aff091ebe}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="TestSpace\Component">
-      <UniqueIdentifier>{c1254fb8-4cbf-453c-b247-9929ef717452}</UniqueIdentifier>
-    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GameProcess.h">
@@ -36,27 +27,8 @@
     <ClInclude Include="ParentScene.h">
       <Filter>Scene</Filter>
     </ClInclude>
-    <ClInclude Include="TestScene.h">
-      <Filter>TestSpace</Filter>
-    </ClInclude>
-    <ClInclude Include="SceneManager.h">
-      <Filter>Scene</Filter>
-    </ClInclude>
-    <ClInclude Include="TestObject_Find.h">
-      <Filter>TestSpace\Object</Filter>
-    </ClInclude>
-    <ClInclude Include="TestObject.h">
-      <Filter>TestSpace\Object</Filter>
-    </ClInclude>
-    <ClInclude Include="TestComponent 1.h">
-      <Filter>TestSpace\Component</Filter>
-    </ClInclude>
-    <ClInclude Include="TestComponent 2.h">
-      <Filter>TestSpace\Component</Filter>
-    </ClInclude>
-    <ClInclude Include="TestComponent.h">
-      <Filter>TestSpace\Component</Filter>
-    </ClInclude>
+    <ClInclude Include="API.h" />
+    <ClInclude Include="importGraphics.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="GameProcess.cpp">
@@ -71,27 +43,7 @@
     <ClCompile Include="ParentScene.cpp">
       <Filter>Scene</Filter>
     </ClCompile>
-    <ClCompile Include="TestScene.cpp">
-      <Filter>TestSpace</Filter>
-    </ClCompile>
-    <ClCompile Include="SceneManager.cpp">
-      <Filter>Scene</Filter>
-    </ClCompile>
-    <ClCompile Include="TestObject_FInd.cpp">
-      <Filter>TestSpace\Object</Filter>
-    </ClCompile>
-    <ClCompile Include="TestObject.cpp">
-      <Filter>TestSpace\Object</Filter>
-    </ClCompile>
-    <ClCompile Include="TestComponent 1.cpp">
-      <Filter>TestSpace\Component</Filter>
-    </ClCompile>
-    <ClCompile Include="TestComponent 2.cpp">
-      <Filter>TestSpace\Component</Filter>
-    </ClCompile>
-    <ClCompile Include="TestComponent.cpp">
-      <Filter>TestSpace\Component</Filter>
-    </ClCompile>
+    <ClCompile Include="API.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="ClassDiagram.cd" />
