+++ Acquitaine Engine/Acquitaine Game Engine/ParentComponent.h
@@ -3,37 +3,41 @@
 
 using namespace std;
 
-class ParentObject;
-
-class ParentComponent
+namespace act
 {
-public:
-	ParentComponent();
-	virtual ~ParentComponent();
 
-//private:
-public:
-	virtual void Initialize();
-	virtual void Finalize();
-	virtual void InputEvent();
-	virtual void Render();
-	virtual void Phsics();
-	//ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ//
+	class ParentObject;
 
-	virtual void Awake();
-	virtual void Enable();
-	virtual void Start();
-	virtual void FixedUpdate();
-	virtual void Update();
+	class ParentComponent
+	{
+	public:
+		ParentComponent();
+		virtual ~ParentComponent();
 
-	virtual void Disable();
-	virtual void Release();
+		//private:
+	public:
+		virtual void Initialize();
+		virtual void Finalize();
+		virtual void InputEvent();
+		virtual void Render();
+		virtual void Phsics();
+		//ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ//
 
-	virtual void SetGameObject(ParentObject* parentObject);
+		virtual void Awake();
+		virtual void Enable();
+		virtual void Start();
+		virtual void FixedUpdate();
+		virtual void Update();
 
-protected:
-	string _componentName;		// 컴포넌트의 이름
-	int _componentID;			// 컴포넌트의 ID.
-	ParentObject* _gameObject;		// 내가 소속된 오브젝트
-	bool _isEnabled;			// 컴포넌트가 활성화 되어 있는지? 아닌지.
-};
+		virtual void Disable();
+		virtual void Release();
+
+		virtual void SetGameObject(ParentObject* parentObject);
+
+	protected:
+		string _componentName;		// 컴포넌트의 이름
+		int _componentID;			// 컴포넌트의 ID.
+		ParentObject* _gameObject;		// 내가 소속된 오브젝트
+		bool _isEnabled;			// 컴포넌트가 활성화 되어 있는지? 아닌지.
+	};
+}
