--- Acquitaine Engine/Acquitaine Game Engine/SceneManager.h
@@ -1,13 +1,0 @@
-#pragma once
-
-class ParentScene;
-class TestScene;
-
-class SceneManager
-{
-public:
-	SceneManager();
-	~SceneManager();
-	void Initialize();
-};
-
