--- Acquitaine Engine/Acquitaine Game Engine/TestObject_FInd.cpp
@@ -1,17 +1,0 @@
-#include "TestObject_Find.h"
-#include "TestComponent.h"
-#include "TestComponent 1.h"
-#include "TestComponent 2.h"
-
-TestObject_Find::TestObject_Find(string objectName, ParentScene* parentScene)
-{
-	this->objectName = objectName;
-	this->_includedScene = parentScene;
-
-	AddComponent<TestComponent2>("component_Find", this);
-}
-
-TestObject_Find::~TestObject_Find()
-{
-
-}
