--- Acquitaine Engine/Acquitaine Game Engine/TestScene.cpp
@@ -1,17 +1,0 @@
-#include "TestScene.h"
-#include "TestObject.h"
-#include "TestObject_Find.h"
-
-TestScene::TestScene(string SceneName)
-{
-	this->_SceneName = SceneName;
-
-	AddObject<TestObject>("test0");
-	AddObject<TestObject>("test1");
-	AddObject<TestObject_Find>("test_find");
-}
-
-TestScene::~TestScene()
-{
-
-}
