--- Acquitaine Engine/Acquitaine Game Engine/TestScene.h
@@ -1,12 +1,0 @@
-#pragma once
-#include "ParentScene.h"
-
-class TestObject;
-
-class TestScene : public ParentScene
-{
-public:
-	TestScene(string SceneName);
-	virtual ~TestScene();
-};
-
